Add getTags function to the CloudProvider Create call
